Voltage Controller Buck and current controller fixes
Made the voltage controller more of a critically damped oscillator (less ringing).

Current controller: Finalized the characteristics of the current controller. Code for actual current controller doesn't work at the moment. Resolution is strange and the jumps in current values is a very large jump so need to fix that issue.